refactor: split equipment into equipment and equipmentClass for organization
